+++ src/main/java/de/budget/project/model/transaction/TransactionWebRequest.java
@@ -2,12 +2,13 @@ package de.budget.project.model.transaction;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
 public class TransactionWebRequest {
     private Date customDate;
-    private Float amount;
+    private BigDecimal amount;
     private Long walletId;
     private Long categoryId;
     private String description;
