+++ src/main/java/de/budget/project/model/transaction/TransactionWebResponse.java
@@ -2,10 +2,12 @@ package de.budget.project.model.transaction;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class TransactionWebResponse {
-    private Float amount;
+    private BigDecimal amount;
     private String description;
     private String categoryName;
-    private Float balance;
+    private BigDecimal balance;
 }
